refactor: levellog 질문 입력 리스트 이름 명확하게 수정
